+++ ft_memcpy.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int				i;
+	size_t				i;
 	unsigned char		*d;
 	const unsigned char	*s;
