+++ ft_memset.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*p;
 
 	i = 0;
